myai-backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
pydantic-settings>=2.2.0
python-dotenv>=1.0.1
httpx>=0.27.0
email-validator>=2.1.0

# Database + migrations
SQLAlchemy[asyncio]>=2.0.30
asyncpg>=0.29.0
alembic>=1.13.1

# Auth
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4


